Add/remove link gui functional
git-svn-id: https://jmg.devguard.com/svn/deops@24 a8feb433-1630-0410-ae02-e0276e0da4ce

+++ Components/Link/LinkControl.cs
@@ -31,7 +31,7 @@ namespace DeOps.Components.Link
 
         internal Dictionary<uint, List<OpLink>> ProjectRoots    = new Dictionary<uint, List<OpLink>>();
         internal Dictionary<ulong, OpLink>      LinkMap         = new Dictionary<ulong, OpLink>();
-        internal Dictionary<ulong, string>      ProjectNames    = new Dictionary<ulong, string>();
+        internal Dictionary<uint, string> ProjectNames = new Dictionary<uint, string>();
 
         Dictionary<ulong, DateTime> NextResearch = new Dictionary<ulong, DateTime>();
         Dictionary<ulong, uint> DownloadLater = new Dictionary<ulong, uint>();
